Fix %d/%u for int32_t/uint32_t on ARM/Cortex-M (long, not int)
On ARM/Cortex-M/Zephyr targets int32_t/uint32_t typedef to long/unsigned long, so %d/%u fail under -Werror=format=. Cast the flagged args to int/unsigned, which is portable regardless of the typedef and matches the cast style already used for adjacent args in these files. Also fix cuda_backend.cpp, where a scoped Error enum was passed to %d.

+++ backends/cuda/runtime/cuda_backend.cpp
@@ -718,7 +718,7 @@ class ET_EXPERIMENTAL CudaBackend final
         error == Error::Ok,
         Internal,
         "AOTInductorModelContainerRun failed with error code %d",
-        error);
+        static_cast<int>(error));
 
     if (is_capture_step) {
       // End capture → instantiate graph
